# internal/api/docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: VoHive 主服务控制面 API — 自动生成 OpenAPI 文档 POC
  title: VoHive Main API
  version: "1.0"
paths:
...
  /devices/{device_id}/network:
    patch:
      consumes:
      - application/json
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceNetworkPatch
      tags:
      - devices
...
  /devices/{device_id}/vowifi:
    patch:
      consumes:
      - application/json
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceVoWiFiPatch
      tags:
      - devices
...
  /proxy-instances/{instance_id}:
    get:
      parameters:
      - description: instance_id
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyInstanceGet
      tags:
      - proxy-instances
  /proxy-instances/{instance_id}/actions/restart:
    post:
      consumes:
      - application/json
      parameters:
      - description: instance_id
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyInstanceRestart
      tags:
      - proxy-instances
  /proxy-instances/{instance_id}/actions/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: instance_id
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyInstanceStart
      tags:
      - proxy-instances
  /proxy-instances/{instance_id}/actions/stop:
    post:
      consumes:
      - application/json
      parameters:
      - description: instance_id
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyInstanceStop
      tags:
      - proxy-instances
  /proxy-instances/config:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyUpdateConfig
      tags:
      - proxy-instances
  /proxy-instances/overview:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ProxyOverview
      tags:
      - proxy-instances
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"